@@ -4,19 +4,21 @@
     "keywords": ["Symfony", "routing", "provider", "modular"],
     "license": "MIT",
     "require": {
-        "php": "^5.6|^7.0",
+        "php": "^7.0",
         "symfony/config": "^2.8|^3.0",
         "symfony/dependency-injection": "^2.8|^3.0",
         "symfony/framework-bundle": "^2.8|^3.0",
         "symfony/http-kernel": "^2.8|^3.0",
-        "symfony/routing": "^2.8|^3.0",
-        "symfony-cmf/routing-bundle": "^1.3"
+        "symfony/routing": "^3.0",
+        "symfony/twig-bundle": "^3.0",
+        "symfony-cmf/routing-bundle": "^1.4"
     },
     "require-dev": {
         "nette/utils": "^2.3",
-        "phpunit/phpunit": "^5.1",
-        "symfony/form": "^2.8|^3.0",
-        "symplify/coding-standard": "^0.1"
+        "phpunit/phpunit": "^5.4",
+        "symfony/form": "^3.1",
+        "symplify/coding-standard": "^0.3",
+        "twig/twig": "^1.24"
     },
     "autoload": {
         "psr-4": {
