@@ -2,16 +2,16 @@
 
 namespace Symplify\ModularRouting\Tests\Routing;
 
-use PHPUnit_Framework_TestCase;
-use Symplify\ModularRouting\Exception\FileNotFoundException;
+use PHPUnit\Framework\TestCase;
 use Symplify\ModularRouting\Tests\Routing\AbstractRouteCollectionProviderSource\MissingFileRouteCollectionProvider;
 
-final class AbstractRouteCollectionProviderTest extends PHPUnit_Framework_TestCase
+final class AbstractRouteCollectionProviderTest extends TestCase
 {
+    /**
+     * @expectedException \Symplify\ModularRouting\Exception\FileNotFoundException
+     */
     public function testMissingFiles()
     {
-        $provider = new MissingFileRouteCollectionProvider();
-        $this->setExpectedException(FileNotFoundException::class);
-        $provider->getRouteCollection();
+        (new MissingFileRouteCollectionProvider())->getRouteCollection();
     }
 }
