@@ -2,6 +2,7 @@
 
 namespace Symplify\ModularRouting\Tests\Routing\AbstractRouteCollectionProviderSource;
 
+use Symfony\Component\Routing\RouteCollection;
 use Symplify\ModularRouting\Routing\AbstractRouteCollectionProvider;
 
 final class FilesRouteCollectionProvider extends AbstractRouteCollectionProvider
@@ -9,7 +10,7 @@ final class FilesRouteCollectionProvider extends AbstractRouteCollectionProvider
     /**
      * {@inheritdoc}
      */
-    public function getRouteCollection()
+    public function getRouteCollection() : RouteCollection
     {
         return $this->loadRouteCollectionFromFiles([
             __DIR__.'/routes.xml',
