@@ -2,6 +2,7 @@
 
 namespace Symplify\ModularRouting\Tests\Routing\AbstractRouteCollectionProviderSource;
 
+use Symfony\Component\Routing\RouteCollection;
 use Symplify\ModularRouting\Routing\AbstractRouteCollectionProvider;
 
 final class MissingFileRouteCollectionProvider extends AbstractRouteCollectionProvider
@@ -9,7 +10,7 @@ final class MissingFileRouteCollectionProvider extends AbstractRouteCollectionPr
     /**
      * {@inheritdoc}
      */
-    public function getRouteCollection()
+    public function getRouteCollection() : RouteCollection
     {
         return $this->loadRouteCollectionFromFile('incorrect-path.yml');
     }
